refactor: separate DimensionParser into parser.ts and export noramlly
default export is not suitable for general purpose
