+++ src/parser/entities/dimension/index.ts
@@ -1,25 +1,2 @@
-import type DxfArrayScanner from '../../DxfArrayScanner';
-import type { ScannerGroup } from '../../DxfArrayScanner';
-import { isMatched } from '../../shared';
-import { parseDimension } from './common';
-import type { DimensionEntity } from './types';
-
-export default class DimensionParser {
-    static ForEntityName = 'DIMENSION';
-
-    parseEntity(scanner: DxfArrayScanner, curr: ScannerGroup) {
-        const entity = {} as DimensionEntity;
-
-        while (!isMatched(curr, 0, 'EOF')) {
-            if (curr.code === 0) {
-                scanner.rewind();
-                return entity;
-            }
-
-            parseDimension(entity, curr, scanner);
-            curr = scanner.next();
-        }
-
-        return entity;
-    }
-}
+export type * from './types'
+export * from './parser'
